KDEGitCommitHooks: Make clang-format hook more interactive
In case the user commits in a terminal we output the diff and
present options to format the files and add all tracked files or just
apply the formatting and manually stage the files afterwards.

In case the user uses not a terminal we output the diff to make it
easier to review the changes. Then the clang-format integration of the
IDE/text editor or the suggested `git clang-format -f` command can be
used to apply the formatting.